backend/requirements.txt
flask
pytesseract~=0.3.13
opencv-python~=4.10.0.84
pillow~=11.0.0
fastapi~=0.115.6
SpeechRecognition~=3.12.0
pyttsx3~=2.98
uvicorn~=0.32.1
numpy~=2.2.0
